[TIMOB-7068] [TIMOB-5511] More work on CommonJS support. Refactored the titanium.js/Ti.js. Added Ti/_/analytics along with XHR support for analytics.
