+++ mobileweb/src/index.html
@@ -33,7 +33,7 @@
 <body>
 	<script>
 	${ti_js}
-	require(['titanium']);
+	require(['Ti']);
 	</script>
 </body>
 </html>
