+++ mobileweb/titanium/Ti/Analytics.js
@@ -1,35 +1,35 @@
-define("Ti/Analytics", ["Ti/_/Evented"], function(Evented) {
-	
-	(function(api){
-		// Interfaces
-		Ti._5.EventDriven(api);
-	
-		// Methods
-		api.addEvent = function(typ, name, data){
-			Ti._5.addAnalyticsEvent(typ, name, data);
-		};
-		api.featureEvent = function(name, data){
-			Ti._5.addAnalyticsEvent('app.feature', name, data);
-		};
-		api.navEvent = function(from, to, name, data){
-			Ti._5.addAnalyticsEvent('app.nav', name, data);
-		};
-		api.settingsEvent = function(name, data){
-			Ti._5.addAnalyticsEvent('app.settings', name, data);
-		};
-		api.timedEvent = function(name, start, stop, duration, data){
-			if(data == null){
-				data = {};
-			}
-			data.start = start;
-			data.stop = stop;
-			data.duration = duration;
-	
-			Ti._5.addAnalyticsEvent('app.timed', name, data);
-		};
-		api.userEvent = function(name, data){
-			Ti._5.addAnalyticsEvent('app.user', name, data);
-		};
-	})(Ti._5.createClass('Ti.Analytics'));
-	
+define(["Ti/_/analytics", "Ti/_/Evented", "Ti/_/lang"], function(analytics, Evented, lang) {
+
+	return lang.setObject("Ti.Analytics", Evented, {
+
+		addEvent: function(type, name, data) {
+			analytics.add(type, name, data);
+		},
+
+		featureEvent: function(name, data) {
+			analytics.add("app.feature", name, data);
+		},
+
+		navEvent: function(from, to, name, data) {
+			analytics.add("app.nav", name, data);
+		},
+
+		settingsEvent: function(name, data) {
+			analytics.add("app.settings", name, data);
+		},
+
+		timedEvent: function(name, start, stop, duration, data) {
+			analytics.add("app.timed", name, require.mix({}, data, {
+				start: start,
+				stop: stop,
+				duration: duration
+			}));
+		},
+
+		userEvent: function(name, data) {
+			analytics.add("app.user", name, data);
+		}
+
+	});
+
 });
