+++ mobileweb/titanium/Ti/App.js
@@ -1,19 +1,7 @@
-define("Ti/App", ["Ti/_/Evented"], function(Evented) {
+define(["Ti/_/Evented", "Ti/_/lang"], function(Evented, lang) {
 
-	(function(api){
-		// Interfaces
-		Ti._5.EventDriven(api);
-	
-		require.mix(api, require.config.app, {
-			idleTimerDisabled: true,
-			proximityDetection: false,
-			proximityState: 0
-		});
-	
-		// Methods
-		api.getArguments = function(){
-			console.debug('Method "Titanium.App.getArguments" is not implemented yet.');
-		};
-	})(Ti._5.createClass('Ti.App'));
+	return lang.mixProps(lang.setObject("Ti.App", Evented), {
+		constants: require.config.app
+	});
 
 });
