+++ mobileweb/titanium/Ti/UI.js
@@ -1,4 +1,4 @@
-define("Ti/UI", ["Ti/_/dom", "Ti/_/Evented", "Ti/_/lang", "Ti/_/ready", "Ti/_/style"], function(dom, Evented, lang, ready, style) {
+define(["Ti/_/dom", "Ti/_/Evented", "Ti/_/lang", "Ti/_/ready", "Ti/_/style"], function(dom, Evented, lang, ready, style) {
 	
 	var body = document.body,
 		isDef = require.isDef,
