+++ mobileweb/titanium/Ti/_/Gestures/GestureRecognizer.js
@@ -1,4 +1,4 @@
-define("Ti/_/Gestures/GestureRecognizer", ["Ti/_/declare", "Ti/_/lang"], function(declare,lang) {
+define(["Ti/_/declare", "Ti/_/lang"], function(declare,lang) {
 
 	return declare("Ti._.Gestures.GestureRecognizer", null, {
