+++ mobileweb/titanium/Ti/_/Gestures/Swipe.js
@@ -1,4 +1,4 @@
-define("Ti/_/Gestures/Swipe", ["Ti/_/declare", "Ti/_/lang","Ti/_/Gestures/GestureRecognizer"], function(declare,lang,GestureRecognizer) {
+define(["Ti/_/declare", "Ti/_/lang","Ti/_/Gestures/GestureRecognizer"], function(declare,lang,GestureRecognizer) {
 
 	return declare("Ti._.Gestures.Swipe", GestureRecognizer, {
