+++ mobileweb/titanium/Ti/_/Layouts/Absolute.js
@@ -1,4 +1,4 @@
-define("Ti/_/Layouts/Absolute", ["Ti/_/Layouts/Base", "Ti/_/declare"], function(Base, declare) {
+define(["Ti/_/Layouts/Base", "Ti/_/declare"], function(Base, declare) {
 
 	return declare("Ti._.Layouts.Absolute", Base, {
