+++ mobileweb/titanium/Ti/_/Layouts/Base.js
@@ -1,4 +1,4 @@
-define("Ti/_/Layouts/Base", ["Ti/_/css", "Ti/_/declare", "Ti/_/style", "Ti/_/dom"], function(css, declare, style, dom) {
+define(["Ti/_/css", "Ti/_/declare", "Ti/_/style", "Ti/_/dom"], function(css, declare, style, dom) {
 	
 	var unitize = dom.unitize,
 		set = style.set,
