+++ mobileweb/titanium/Ti/_/Layouts/Horizontal.js
@@ -1,4 +1,4 @@
-define("Ti/_/Layouts/Horizontal", ["Ti/_/Layouts/Base", "Ti/_/declare"], function(Base, declare) {
+define(["Ti/_/Layouts/Base", "Ti/_/declare"], function(Base, declare) {
 
 	return declare("Ti._.Layouts.Horizontal", Base, {
