+++ mobileweb/titanium/Ti/_/UI/Widget.js
@@ -1,4 +1,4 @@
-define("Ti/_/UI/Widget", ["Ti/_/declare", "Ti/UI/View"], function(declare, View) {
+define(["Ti/_/declare", "Ti/UI/View"], function(declare, View) {
 
 	// base class for various widgets that will eventually merge with Ti._.UI.Element in 1.9
 	return declare("Ti._.UI.Widget", View);
