+++ mobileweb/titanium/Ti/_/browser.js
@@ -1,4 +1,4 @@
-define("Ti/_/browser", function() {
+define(function() {
 	var match = navigator.userAgent.toLowerCase().match(/(webkit|gecko|trident|presto)/);
 	return {
 		runtime: match ? match[0] : "unknown"
